--- src/components/NewExpense/ExpensesFilter.css
@@ -1,35 +1,0 @@
-.expenses-filter__box {
-  background-color: #a892ee;
-  padding: 1rem;
-  margin: 2rem auto;
-  width: 50rem;
-  max-width: 95%;
-  border-radius: 12px;
-  text-align: center;
-  box-shadow: 0 1px 8px rgba(0, 0, 0, 0.25);
-}
-
-.expenses-filter {
-  color: white;
-  padding: 0 1rem;
-}
-
-.expenses-filter__control {
-  display: flex;
-  width: 100%;
-  align-items: center;
-  justify-content: space-between;
-  margin: 1rem 0;
-}
-
-.expenses-filter label {
-  font-weight: bold;
-  margin-bottom: 0.5rem;
-}
-
-.expenses-filter select {
-  font: inherit;
-  padding: 0.5rem 3rem;
-  font-weight: bold;
-  border-radius: 6px;
-}
